requirements.txt
# RL core
gymnasium[classic_control,atari]        # entornos básicos y Atari
gym-minigrid>=1.0.2                     # MiniGrid para grid-world simbólico
procgen==0.10.5                         # Procgen Benchmark

# Deep learning
torch>=2.0.0,<3.0                       # PyTorch para redes y entrenamiento
torchvision>=0.15.0                     # utilidades de imagen

# Agentes de referencia
stable-baselines3[extra]>=2.1.0         # SB3 con dependencias extra (Atari, Tensorboard)
ray[rllib]>=2.5.0                       # RLlib para comparar con DQN/PPO escalable

# Utilidades y logging
numpy>=1.24.0                           # arrays y procesamiento numérico
pandas>=2.0.0                           # análisis de resultados y CSV
matplotlib>=3.7.0                       # gráficas de curvas de aprendizaje
tensorboard>=2.12.0                     # visualización de entrenamiento
wandb>=0.15.0                           # (opcional) experiment tracking

# Extras recomendados
tqdm>=4.65.0                            # barras de progreso
pytest>=7.4.0                           # tests unitarios de tu código
black                                 # formateo de código (opcional)
